requirements.txt
ultralytics
streamlit
Pillow
